# openapi.yaml
openapi: 3.0.3
info:
  title: Sistem Pendataan Inventaris Laboratorium REST API
  description: |
    API Spesifikasi untuk Sistem Pendataan Inventaris Laboratorium (Sistem Pendataan Inventaris Laboratorium) 
    menggunakan Spring Boot, JWT Authentication, dan Role-Based Access Control (RBAC).
    
    ### Peran Pengguna (Roles):
    - **ADMIN**: Memiliki akses penuh ke seluruh sistem termasuk CRUD kategori, barang, laporan, dan aksi peminjaman.
    - **PETUGAS**: Memiliki hak akses serupa dengan ADMIN untuk operasional sehari-hari (CRUD kategori/barang, persetujuan peminjaman, laporan).
    - **MAHASISWA**: Hanya dapat melihat kategori/barang, melakukan pengajuan peminjaman barang untuk diri sendiri, dan melihat riwayat peminjaman miliknya sendiri.
    
    ### Alur Autentikasi:
    1. Lakukan registrasi melalui `/api/auth/register` (opsional jika menggunakan akun seeder bawaan).
    2. Lakukan login melalui `/api/auth/login` untuk mendapatkan token JWT.
    3. Masukkan token tersebut dalam Header HTTP: `Authorization: Bearer <JWT_TOKEN>` untuk mengakses endpoint yang dilindungi.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Development Server
paths:
...
  /api/kategori:
    get:
      tags:
        - Kategori
      summary: Dapatkan Semua Kategori
      description: Mengambil daftar kategori barang inventaris dengan paginasi, pengurutan, dan pencarian data. Dapat diakses oleh ADMIN, PETUGAS, dan MAHASISWA.
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Nomor halaman (0-indexed)
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Jumlah data per halaman
          required: false
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          description: Kolom/properti untuk pengurutan
          required: false
          schema:
            type: string
            default: idKategori
        - name: sortDir
          in: query
          description: Arah pengurutan (ASC atau DESC)
          required: false
          schema:
            type: string
            enum: [ASC, DESC]
            default: ASC
      responses:
        '200':
          description: Berhasil memuat daftar kategori
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseListKategoriResponse'
        '401':
          description: Token tidak valid atau tidak disertakan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseString'

    post:
      tags:
        - Kategori
      summary: Buat Kategori Baru
      description: Menambahkan kategori barang baru ke sistem. Hanya dapat diakses oleh **ADMIN** dan **PETUGAS**.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KategoriRequest'
      responses:
        '201':
          description: Kategori berhasil dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseKategoriResponse'
        '400':
          description: Validasi gagal / Nama kategori sudah digunakan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseValidationError'
        '403':
          description: Hak akses ditolak (Hanya ADMIN/PETUGAS)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseString'
  /api/kategori/{id}:
    get:
      tags:
        - Kategori
      summary: Detail Kategori berdasarkan ID
      description: Mengambil informasi detail satu kategori barang. Dapat diakses oleh ADMIN, PETUGAS, dan MAHASISWA.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: ID Kategori
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Kategori ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseKategoriResponse'
        '404':
          description: Kategori tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseString'

    put:
      tags:
        - Kategori
      summary: Perbarui Kategori
      description: Memperbarui data kategori berdasarkan ID. Hanya dapat diakses oleh **ADMIN** dan **PETUGAS**.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: ID Kategori yang akan diperbarui
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KategoriRequest'
      responses:
        '200':
          description: Kategori berhasil diperbarui
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseKategoriResponse'
        '404':
          description: Kategori tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseString'
        '400':
          description: Validasi gagal / Nama kategori duplikat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseValidationError'
        '403':
          description: Hak akses ditolak

    delete:
      tags:
        - Kategori
      summary: Hapus Kategori
      description: Menghapus kategori berdasarkan ID. Hanya dapat diakses oleh **ADMIN** dan **PETUGAS**.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: ID Kategori yang akan dihapus
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Kategori berhasil dihapus
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseString'
        '404':
          description: Kategori tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseString'
        '403':
          description: Hak akses ditolak
...
  /api/peminjaman:
    get:
      tags:
        - Transaksi Peminjaman
      summary: Dapatkan Riwayat/Daftar Transaksi Peminjaman
      description: |
        Mengambil daftar pengajuan peminjaman barang dengan paginasi.
        - **ADMIN** dan **PETUGAS** dapat melihat seluruh transaksi dari semua pengguna.
        - **MAHASISWA** hanya dapat melihat riwayat peminjaman miliknya sendiri.
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Nomor halaman (0-indexed)
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Jumlah data per halaman
          required: false
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          description: Kolom untuk pengurutan
          required: false
          schema:
            type: string
            default: idPeminjaman
        - name: sortDir
          in: query
          description: Arah pengurutan (ASC atau DESC)
          required: false
          schema:
            type: string
            enum: [ASC, DESC]
            default: DESC
      responses:
        '200':
          description: Berhasil memuat daftar transaksi peminjaman
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseListPeminjamanResponse'
        '401':
          description: Token tidak valid

    post:
      tags:
        - Transaksi Peminjaman
      summary: Ajukan Peminjaman Barang
      description: |
        Mengajukan peminjaman satu atau beberapa barang inventaris sekaligus.
        - Status peminjaman default diset ke `MENUNGGU` (Pending).
        - Sistem belum memotong stok saat pengajuan dibuat.
        - Tanggal pinjam harus hari ini atau di masa mendatang.
        - Dapat diakses oleh ADMIN, PETUGAS, dan MAHASISWA.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PeminjamanRequest'
      responses:
        '201':
          description: Pengajuan peminjaman berhasil dibuat, menunggu persetujuan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponsePeminjamanResponse'
        '400':
          description: Validasi gagal / Tanggal pinjam melanggar aturan / ID barang salah
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseValidationError'
...
  /api/laporan/stok:
    get:
      tags:
        - Laporan Ringkasan
      summary: Unduh Laporan Stok Inventaris
      description: |
        Menyediakan data ringkasan metrik statistik stok laboratorium yang aktual:
        - Jumlah total kategori, unit barang, dan total stok sistem.
        - Total stok yang saat ini tersedia dan sedang dipinjam.
        - Distribusi kondisi barang (BAIK, RUSAK_RINGAN, RUSAK_BERAT).
        Hanya dapat diakses oleh **ADMIN** dan **PETUGAS**.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Laporan stok inventaris berhasil dimuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseLaporanStokResponse'
        '401':
          description: Token tidak valid
        '403':
          description: Hak akses ditolak
  /api/laporan/peminjaman:
    get:
      tags:
        - Laporan Ringkasan
      summary: Laporan Statistik Peminjaman
      description: |
        Menampilkan data ringkasan total transaksi peminjaman secara kumulatif serta jumlah transaksi berdasarkan statusnya 
        (Menunggu, Disetujui, Ditolak, Dikembalikan).
        Hanya dapat diakses oleh **ADMIN** dan **PETUGAS**.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Laporan transaksi peminjaman berhasil dimuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponseLaporanPeminjamanResponse'
        '401':
          description: Token tidak valid
        '403':
          description: Hak akses ditolak
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Masukkan token JWT Anda dalam format `Bearer <token>` pada header Otorisasi HTTP.
  schemas:
...
    KategoriRequest:
      type: object
      required:
        - namaKategori
      properties:
        namaKategori:
          type: string
          maxLength: 100
          example: Alat Ukur Elektronik
        deskripsi:
          type: string
          example: Komponen alat ukur elektronik praktikum teknik elektro (Multimeter, Osciloscope, Signal Generator, dll)
...
    PeminjamanRequest:
      type: object
      required:
        - tglPinjam
        - tglKembali
        - details
      properties:
        tglPinjam:
          type: string
          format: date
          example: "2026-05-18"
          description: Tanggal peminjaman barang (harus hari ini atau masa mendatang)
        tglKembali:
          type: string
          format: date
          example: "2026-05-25"
          description: Tanggal rencana pengembalian barang
        catatan:
          type: string
          example: Untuk keperluan praktikum mata kuliah Mikroprosesor minggu ini.
        details:
          type: array
          items:
            $ref: '#/components/schemas/DetailPeminjamanRequest'
    DetailPeminjamanRequest:
      type: object
      required:
        - idBarang
        - jumlah
      properties:
        idBarang:
          type: integer
          format: int64
          example: 1
        jumlah:
          type: integer
          minimum: 1
          example: 2
...
    Role:
      type: string
      enum:
        - ADMIN
        - PETUGAS
        - MAHASISWA
      example: ADMIN
    KondisiBarang:
      type: string
      enum:
        - BAIK
        - RUSAK_RINGAN
        - RUSAK_BERAT
      example: BAIK
    StatusPeminjaman:
      type: string
      enum:
        - MENUNGGU
        - DISETUJUI
        - DITOLAK
        - DIKEMBALIKAN
      example: MENUNGGU

    # Wrapper Response Objects
    WebResponseString:
      type: object
      properties:
        code:
          type: integer
          example: 200
        status:
          type: string
          example: OK
        message:
          type: string
          example: Operasi berhasil!
        data:
          type: string
          nullable: true
          example: null
    WebResponseValidationError:
      type: object
      properties:
        code:
          type: integer
          example: 400
        status:
          type: string
          example: BAD REQUEST
        message:
          type: string
          example: Validation failed
        data:
          type: object
          additionalProperties:
            type: string
          example:
            email: Format email tidak valid!
            password: Password minimal terdiri dari 6 karakter!
    PagingResponse:
      type: object
      properties:
        currentPage:
          type: integer
          example: 0
        totalPages:
          type: integer
          example: 1
        size:
          type: integer
          example: 10
        totalElements:
          type: integer
          format: int64
          example: 5

    # Entity Response Objects
    UserResponse:
      type: object
      properties:
        idUser:
          type: integer
          format: int64
          example: 1
        nama:
          type: string
          example: Ahmad Mahasiswa
        email:
          type: string
          format: email
          example: mahasiswa@lab.com
        role:
          $ref: '#/components/schemas/Role'
        createdAt:
          type: string
          format: date-time
          example: "2026-05-18T08:50:00Z"
...
    KategoriResponse:
      type: object
      properties:
        idKategori:
          type: integer
          format: int64
          example: 1
        namaKategori:
          type: string
          example: Alat Ukur Elektronik
        deskripsi:
          type: string
          example: Komponen alat ukur elektronik praktikum teknik elektro (Multimeter, Osciloscope, Signal Generator, dll)
    WebResponseKategoriResponse:
      type: object
      properties:
        code:
          type: integer
          example: 201
        status:
          type: string
          example: CREATED
        message:
          type: string
          example: Kategori berhasil dibuat!
        data:
          $ref: '#/components/schemas/KategoriResponse'
    WebResponseListKategoriResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        status:
          type: string
          example: OK
        message:
          type: string
          example: Berhasil mendapatkan daftar kategori!
        data:
          type: array
          items:
            $ref: '#/components/schemas/KategoriResponse'
        paging:
          $ref: '#/components/schemas/PagingResponse'
...
    DetailPeminjamanResponse:
      type: object
      properties:
        idDetail:
          type: integer
          format: int64
          example: 1
        idBarang:
          type: integer
          format: int64
          example: 1
        namaBarang:
          type: string
          example: Digital Multimeter Sanwa CD800a
        kodeBarang:
          type: string
          example: BRG-EL-001
        jumlah:
          type: integer
          example: 2
        kondisiKembali:
          anyOf:
            - $ref: '#/components/schemas/KondisiBarang'
            - type: string
              nullable: true
          description: Kondisi barang saat dikembalikan. Berstatus `null` selama status pinjam belum DIKEMBALIKAN.
    PeminjamanResponse:
      type: object
      properties:
        idPeminjaman:
          type: integer
          format: int64
          example: 1
        peminjam:
          $ref: '#/components/schemas/UserResponse'
        petugas:
          anyOf:
            - $ref: '#/components/schemas/UserResponse'
            - type: object
              nullable: true
          description: Petugas/Admin yang melakukan verifikasi transaksi. Bernilai `null` jika status peminjaman masih MENUNGGU.
        tglPinjam:
          type: string
          format: date
          example: "2026-05-18"
        tglKembali:
          type: string
          format: date
          example: "2026-05-25"
        status:
          $ref: '#/components/schemas/StatusPeminjaman'
        catatan:
          type: string
          example: Untuk keperluan praktikum mata kuliah Mikroprosesor minggu ini.
        details:
          type: array
          items:
            $ref: '#/components/schemas/DetailPeminjamanResponse'
    WebResponsePeminjamanResponse:
      type: object
      properties:
        code:
          type: integer
          example: 201
        status:
          type: string
          example: CREATED
        message:
          type: string
          example: Pengajuan peminjaman berhasil dibuat, menunggu persetujuan petugas!
        data:
          $ref: '#/components/schemas/PeminjamanResponse'
    WebResponseListPeminjamanResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        status:
          type: string
          example: OK
        message:
          type: string
          example: Daftar transaksi peminjaman berhasil dimuat!
        data:
          type: array
          items:
            $ref: '#/components/schemas/PeminjamanResponse'
        paging:
          $ref: '#/components/schemas/PagingResponse'
    LaporanStokResponse:
      type: object
      properties:
        totalKategori:
          type: integer
          format: int64
          example: 5
        totalBarangUnit:
          type: integer
          format: int64
          example: 25
        totalStokSistem:
          type: integer
          format: int64
          example: 230
        totalStokTersedia:
          type: integer
          format: int64
          example: 215
        totalStokDipinjam:
          type: integer
          format: int64
          example: 15
        barangKondisiBaik:
          type: integer
          format: int64
          example: 220
        barangKondisiRusakRingan:
          type: integer
          format: int64
          example: 8
        barangKondisiRusakBerat:
          type: integer
          format: int64
          example: 2
    WebResponseLaporanStokResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        status:
          type: string
          example: OK
        message:
          type: string
          example: Laporan stok inventaris berhasil dimuat!
        data:
          $ref: '#/components/schemas/LaporanStokResponse'
    LaporanPeminjamanResponse:
      type: object
      properties:
        totalPeminjaman:
          type: integer
          format: int64
          example: 45
        statusMenunggu:
          type: integer
          format: int64
          example: 3
        statusDisetujui:
          type: integer
          format: int64
          example: 12
        statusDitolak:
          type: integer
          format: int64
          example: 5
        statusDikembalikan:
          type: integer
          format: int64
          example: 25
    WebResponseLaporanPeminjamanResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        status:
          type: string
          example: OK
        message:
          type: string
          example: Laporan transaksi peminjaman berhasil dimuat!
        data:
          $ref: '#/components/schemas/LaporanPeminjamanResponse'